Calculate the is_blocked flag correctly
This changes the BLOCKED label to be based on the new 'is_blocked' flag.
The value of that flag is calculated by looking at all the bugs the bug
depends on and if any are open, then the 'is_blocked' flag is set to
True.

This does some extra bookkeeping in order to avoid an extra Bugzilla API
call, but does at most one additional call.

Fixes #4