Nix cache in favor of Flask sessions
The minor complexity is that Flask sessions are HttpOnly, so Angular
can't see them. So we have to store things that Angular needs to know in
"regular cookies" and not the Flask session.

Otherwise this nixes all the cache infrastructure in exchange for
storing things in Flask sessions which are more resilient to
starting/stopping the server when using SimpleCache and MemCache cachy
expirationy things.