Add star; rewrite "show/hide closed"
This adds the ability to "star" bugs. Starring a bug stores that id in
local storage. The thing this makes possible is you can star a few
things in a sprint and then only focus on those things rather than
having to see the entirety of the sprint which is somewhat exhausting.

Further, I rewrote the "show/hide closed" stuff to use a filter.

Fixes #22
Fixes #45