Add notion of "jeopardy" to sprint details page
If a bug is "in trouble", then it'll show up in yellow and then red
depending on how "in trouble" it is in regards to the estimated point
value and the sprint end date.

This hard-codes the notion of points -> hours. That should get moved to
the sprint metadata in the db.

Fixes #59