Add github PRs to ernest sprint details view
If a bug has a related PR, this will show the PR with a link. The way to
denote the bug for the PR is to have something like:

    [bug xyz]

in the PR summary. This supports multiple bugs. For example:

    [bug xyz] [bug abc]

This also adds some project admin stuff to add the github owner and
repo data to the Project table.

Fixes #60