Fix sprint details view regarding sprint searches
If a sprint name was a prefix for other sprint names (e.g. 2014.1 is a
prefix of 2014.10), then we'd pull all the bugs for all the sprints with
that prefix. Ew.

This fixes that by switching to a regexp search for the whiteboard field
and appending a [[:>:]] to the end which is an empty character that says
"the next thing can't be a word character".

Fixes #63