Stopgap fix for autorefresh issues
If you're editing the notes on the sprint details page and the
autorefresh kicks in, then you lose all the things you wrote so far.
That sucks.

This is a stopgap fix in that it's not a great fix, but it's easy to do
and it alleviates the problem. It adds a button letting you
enable/disable the autorefresh. It also changes the page so that it
defaults to autorefresh being disabled.

Fixes #75